Switched imu_angle to be a float like the imu driver produces

+++ teensy_chibios/src/main/include/imu.h
@@ -1,7 +1,7 @@
 #ifndef IMU_H
 #define IMU_H
 
-short imu_loop_fn();
+float imu_loop_fn();
 
 void imu_setup();
